buildConfig to show version number
Build Config was added to show dynamic version name / number
